Implement backend of fuel system (for mandatory stops right now)

+++ Assets/Scripts/PersistentManager/PersistentManager.cs
@@ -29,7 +29,7 @@ namespace PersistentManager
 
         #region Train Data
 
-        [Header("Train Data")] public float Fuel;
+        [Header("Train Data")] public int Fuel;
         public TravelCostData TravelCostData;
         public int RouteProgress;
         public PossibleStop CurrentStop;
